feat: fix dependencies

+++ android/src/main/java/tech/notifly/rn/NotiflySdkModule.kt
@@ -130,9 +130,9 @@ class NotiflySdkModule internal constructor(private val reactContext: ReactAppli
   }
 
   @ReactMethod
-  override fun setLogLevel(logLevel: Int, promise: Promise) {
+  override fun setLogLevel(logLevel: Double, promise: Promise) {
     try {
-      Notifly.setLogLevel(logLevel)
+      Notifly.setLogLevel(logLevel.toInt())
       promise.resolve(null)
     } catch (e: Exception) {
       promise.reject(e)
